Move/rename auto script infrastructure to make it more generic
Now just called "ScriptedCommand"
